Update for code review
