+++ src/utils/interaction.js
@@ -1,7 +1,4 @@
-/* eslint-disable import/no-cycle */
-import { items, render } from './structure.js';
-
-function textDecoration(listInput) {
+const textDecoration = (listInput) => {
   listInput.forEach((item) => {
     if (item.hasAttribute('checked')) {
       item.nextSibling.style.textDecoration = 'line-through';
@@ -9,9 +6,9 @@ function textDecoration(listInput) {
       item.nextSibling.style.textDecoration = 'none';
     }
   });
-}
+};
 
-function userInteraction(listInput) {
+const userInteraction = (listInput) => {
   const itemsLocal = JSON.parse(localStorage.getItem('itemsLocal'));
   listInput.forEach((item) => {
     item.addEventListener('change', () => {
@@ -30,36 +27,8 @@ function userInteraction(listInput) {
       }
       textDecoration(listInput);
       localStorage.setItem('itemsLocal', JSON.stringify(itemsLocal));
-      items.splice(0, items.length, ...itemsLocal);
     });
   });
-}
+};
 
-function storeValues() {
-  document.querySelectorAll('.list-input').forEach((item) => {
-    item.addEventListener('change', () => {
-      localStorage.setItem('itemsLocal', JSON.stringify(items));
-    });
-  });
-}
-
-function populateStorage() {
-  window.addEventListener('load', () => {
-    render();
-    const listInput = document.querySelectorAll('.list-input');
-    const itemsLocal = JSON.parse(localStorage.getItem('itemsLocal'));
-    listInput.forEach((item) => {
-      const parent = item.parentNode;
-      const superParent = parent.parentNode;
-      const index = Array.prototype.indexOf.call(superParent.children, parent);
-      const currentItem = itemsLocal[index].completed;
-      if (currentItem) {
-        item.setAttribute('checked', '');
-        parent.lastChild.style.display = 'block';
-      }
-    });
-    textDecoration(listInput);
-  });
-}
-
-export { storeValues, populateStorage, userInteraction };
+export { userInteraction };
